% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{center}
\alias{center}
\title{Center}
\usage{
center(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions to return; defaults to
no dimensionality reduction if \code{k} is \code{NULL}}
}
\value{
Centered \code{m}
}
\description{
Centers \code{m}
}

