% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encoding.data.frame}
\alias{encoding.data.frame}
\title{Create \code{encoding} object}
\usage{
\method{encoding}{data.frame}(d, transformation = NULL, ...)
}
\arguments{
\item{d}{a \code{\link{data.frame}}}

\item{transformation}{a transformation that will be applied to \code{m}}
}
\value{
See \code{\link{encoding}}
}
\description{
Create \code{encoding} object from \code{\link{data.frame}}
with \code{label} corresponding to a column called \code{label} in \code{d}, \code{m}
corresponding to the remaining columns, and \code{fname} corresponding to the
names of the remaining columns
}

