% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encoding}
\alias{encoding}
\title{Create \code{encoding} object}
\usage{
encoding(...)
}
\value{
An object of class \code{encoding} with three elements: \code{m},
an \code{n} by \code{k} matrix of \code{n} objects coded in \code{k} features;
\code{label}, a string vector labelling the objects of \code{m}; \code{fname},
a string vector labelling the features of \code{m}
}
\description{
Generic method to create an \code{encoding} object
}

