% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptest.R
\name{joint_mptests}
\alias{joint_mptests}
\title{Conduct minimal pair tests}
\usage{
joint_mptests(encoding, test_pairs, similarity)
}
\arguments{
\item{encoding}{an \code{\link{encoding}} object}

\item{test_pairs}{a data frame specifying the minimal pair tests: it contains a column
\code{fname} giving the feature that each pair is a minimal pair for, a column \code{x1}
giving the label of the left element of the pair, and a column \code{x2} giving the label 
of the right element of the pair}

\item{similarity}{a similarity function to apply to encoded pairs}
}
\value{
a \code{\link{dplyr::tbl}} containing the results of minimal pair tests for all the
unique feature names in \code{test_pairs$fname} that are not \code{NA}; each feature is
tested by comparing all its pairs (all the pairs labelled with that feature name) to
each other, on the one hand, and to all the other pairs in \code{test_pairs}
(including pairs whose labels are \code{NA}), on the other; encodings of pairs will be compared
by \code{similarity}, where the encoding of a pair is \code{x1-x2}; the \code{tbl} will
have the following columns:
\itemize{
 \itemize{"fname"}{feature name}
 \itemize{"data"}{a nested \code{tbl} containing all the pairs that were taken in both the "same" (both \code{fname}) and
 "different" (only one \code{fname}) groups}
 \itemize{"roc"}{a nested \code{tbl} containing a ROC curve (three columns: \code{fpr}, false positive rate;
 \code{tpr}, the highest corresponding true positive rate; \code{crit}, the highest corresponding linear decision
 boundary)}
 \itemize{"auc"}{AUC (integral under \code{roc$tpr ~ roc$fpr})}
}
}
\description{
Conducts minimal pair tests jointly on the features defined in
\code{test_pairs}
}

