% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reduce_to}
\alias{reduce_to}
\title{Reduce number of dimensions}
\usage{
reduce_to(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions to return}
}
\value{
The first \code{k} columns of \code{m}, or \code{m} if \code{k} is \code{NULL}
or larger than the number of columns of \code{m} - 1
}
\description{
reduces to a fixed number of dimensions \code{k}, or does nothing if
\code{k} is \code{NULL} or larger than the number of columns of \code{m} - 1
}

